Django/restShop/requirements.txt
django>=2.0.4
djangorestframework>=3.8.2
mysqlclient>=1.3.12
markdown>=2.6.11
pillow>=5.1.0
django-crispy-forms>=1.7.2
django-import-export>=1.0.0
django-reversion>=2.0.0
django-formtools>=2.1
future>=0.16.0
httplib2>=0.11.3
six>=1.11.0
xlsxwriter>=1.0.4
xlwt
django-guardian>=1.4.9
coreapi